add confirmation upon suppression of modules and files.

+++ src/app/modules/modules-manager/modules-manager.component.ts
@@ -49,8 +49,11 @@ export class ModulesManagerComponent implements OnInit, OnDestroy {
   }
 
   deleteModule(moduleId: number) {
-    this.moduleService.deleteModule(moduleId);
-    this.modules = this.modules.filter( module => module.id != moduleId);
+    // TODO: replace confirm by a beautiful modal
+    if (confirm("Supprimer le module ?")){
+      this.moduleService.deleteModule(moduleId);
+      this.modules = this.modules.filter( module => module.id != moduleId);
+    }
   }
 
   openModal(template: TemplateRef<any>) {
